refactor: consolidate tap AUT frame files under lib/tap/aut
Move aut-frame, frame-cdp, and frame-scripts into a single lib/tap/aut/
folder as frame.ts, cdp.ts, and scripts.ts, dropping the redundant prefixes.
Update the frame command imports and specs accordingly.
