+++ cli/lib/tap/aut/scripts.ts
